refactor: NavController StateHosting
